fix golint for pkg/volume/azure_dd
