+++ cmd/kube-controller-manager/app/BUILD
@@ -87,8 +87,8 @@ go_library(
         "//pkg/serviceaccount:go_default_library",
         "//pkg/volume:go_default_library",
         "//pkg/volume/awsebs:go_default_library",
-        "//pkg/volume/azure_dd:go_default_library",
         "//pkg/volume/azure_file:go_default_library",
+        "//pkg/volume/azuredd:go_default_library",
         "//pkg/volume/cinder:go_default_library",
         "//pkg/volume/csi:go_default_library",
         "//pkg/volume/csimigration:go_default_library",
