+++ hack/.golint_failures
@@ -185,7 +185,6 @@ pkg/util/removeall
 pkg/util/selinux
 pkg/util/taints
 pkg/volume
-pkg/volume/azure_dd
 pkg/volume/azure_file
 pkg/volume/csi/fake
 pkg/volume/git_repo
