+++ pkg/kubelet/BUILD
@@ -220,7 +220,7 @@ go_test(
         "//pkg/util/taints:go_default_library",
         "//pkg/volume:go_default_library",
         "//pkg/volume/awsebs:go_default_library",
-        "//pkg/volume/azure_dd:go_default_library",
+        "//pkg/volume/azuredd:go_default_library",
         "//pkg/volume/gcepd:go_default_library",
         "//pkg/volume/hostpath:go_default_library",
         "//pkg/volume/testing:go_default_library",
