+++ pkg/volume/azuredd/azure_provision.go
@@ -16,7 +16,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package azure_dd
+package azuredd
 
 import (
 	"errors"
